git add README.md
MusicDex is a platform built on Solana that makes entering the blockchain world simple and intuitive through music. We allow artists to tokenize their tracks and investors to earn passive income from royalties on platforms like Spotify and Apple Music. Thanks to MusicDex, around 1,000–1,500 people have already started using Solana, many of them for the first time. We are actively expanding this segment to grow our user base, and we are developing a mobile game that will first be available to play on Telegram, and later on Google Play and the App Store. This will help us attract even more new users to the Solana ecosystem.