Update README.md
Right now, the legal process is underway — we’re officially registering everything, while the IT team is already uploading the backend for the game. In just a few weeks, it’ll be ready for testing. This stage brings us even closer to launching the full version and attracting thousands of new Solana users through the app